#0000: Fixed sign extensions TESTS_RAN: Unit test, dev test
git-svn-id: https://10.196.60.217/svn/schooner-products/branches/fdf-3.0@2515 01a69087-22d5-4a29-8152-8a1d5e10e5e9

+++ ssd/fifo/mcd_rec.h
@@ -453,10 +453,10 @@ typedef struct mcd_rec_pp_state {
     int                     fill_count;    // recs in logbuf added to pp list
     int                     slot_count;    // slots in logbuf already pp'd
     int                     dealloc_count; // number of dealloc block addresses
-    uint32_t              * dealloc_list;  // list of dealloc block addresses
+    uint64_t              * dealloc_list;  // list of dealloc block addresses
     uint                    dealloc_head,  // ring buffer to delay slab reuse
                             dealloc_tail;  //          by a guaranteed amount
-    uint32_t                dealloc_ring[80000];
+    uint64_t                dealloc_ring[80000];
 	bool                    dealloc_ring_enabled;
 } mcd_rec_pp_state_t;
